Make connections closeable
When being canceled, the connection's run loop is usually stuck trying
to read from the network connection. We make it return immediately by
closing the connection FD, but this returns an error which we don't
want to return. We use Go's context.WithCancelCause() in order to tell
whether this was an intentional shutdown or not. (Using a regular
context.Err() would also work, but there would be a chance of a race
condition when doing that, if the context gets canceled independently
of the error.)